requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
azure-storage-blob==12.19.0
python-dotenv==1.0.1
werkzeug==3.0.1
pyodbc==5.1.0
